Siirretty staticmethodit /commands/ alle
