+++ src/commands/poyta.py
@@ -1,0 +1,6 @@
+from telegram.ext import CallbackContext
+from telegram import Update
+import config
+
+def poyta(update: Update, context: CallbackContext):
+    context.bot.send_animation(chat_id=update.message.chat_id, animation=config.desk, disable_notification=True)
